11 10 11:43

+++ src/main/java/com/ds/watchtable/repository/PosRepository.java
@@ -2,6 +2,10 @@ package com.ds.watchtable.repository;
 
 import com.ds.watchtable.entity.Pos;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface PosRepository extends JpaRepository<Pos, Integer> {
+import java.util.List;
+
+public interface PosRepository extends JpaRepository<Pos, Long> {
+//    List<Object[]> getPosAll(Long posNum);
 }
